doed-messenger/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
sqlalchemy==1.4.52
python-dotenv==1.0.0
argon2-cffi==23.1.0